Add Camera value view

+++ app/src/main/java/com/chaeyoon/haebit/obscura/core/Camera.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface Camera {
     val aperture: Float
-    val isoFlow: StateFlow<Int>
+    val isoFlow: StateFlow<Float>
     val shutterSpeedFlow: StateFlow<Float>
     val exposureValueFlow: StateFlow<Float>
